Delete the license, as I'm not the game creator
Paul Baldowski, creator of The Dee Sanction RPG, will provide a charactersheet for the game for Roll20 once the game is complete and published. Until then it would be cheeky of me to attach an arbitrary license to my version of his draft charactersheet. So I've deleted the one I chose more or less at random while creating this repo.